Fix workaround for get requests.
Basically the workaround is a wrapper for get requests where it will
loop a couple of times, sleeping for one minute between each set of
attempts, which is necessary in case the server shuts down for a bit.

For some reason there is an issue that I cannot reproduce (maybe it is
Windows only) that making this wrapper actually work might fix.

Also formatting.