Increase maximum depth for get request wrapper
Sometimes the server can go down for longer than usual (3 minutes
instead of 30 seconds), so it's better to wait a couple more times in
case they don't have a cache.

Ideally this will be changed eventually to have a dynamic wait time,
starting at 30 seconds and incrementing over the course of 3 or 4
attempts to a total wait of around 5 minutes, but this is out of the
scope of this specific branch and overall not necessary.