"Fix" issue with networking on windows (#2)
Fixes get wrapper and WinError 10060

The get wrapper is meant to wait for the server to go back up if it goes
down for ~5 mins or less, but it was catching the wrong error and
therefore wasn't working. This also apparently serves as a workaround
for WinError 10060 (by just waiting a bit before trying to request
again).